Conditionally look for aws_acm_certificate based on cert_domain var
If the varialbe cert_domain is set, then don't attempt to find a data
source that corresponds to the aws_acm_certificate. This change allows
you to create the aws_acm_certificate alongside of this module in
Terraform without receiving a "No certificate for domain found in this
region" error. This change also has no effect on the other code path,
where you leave cert_domain unset and instead rely on the data source.